Rename adapter/aisdk to adapter/vercelaisdk/uimessagestream
The old name was confusing — "aisdk" inside "ai-sdk-go" says nothing.
The new structure makes the hierarchy clear:

  adapter/vercelaisdk/uimessagestream/  — UI Message Stream protocol
  adapter/a2a/                          — Agent-to-Agent protocol

Room to add more Vercel AI SDK adapters later (text stream, completion)
under the vercelaisdk/ namespace.
