+++ adapter/vercelaisdk/uimessagestream/conformance/.gitignore
