+++ adapter/vercelaisdk/uimessagestream/conformance/package.json
