+++ adapter/vercelaisdk/uimessagestream/conformance/run.sh
